Started working on arbitrary Expressions in test steps
