+++ Engine/MenuAnnotation.cs
@@ -197,5 +197,7 @@ namespace OpenTap
         public const string AddDynamicProperty = Common + nameof(AddDynamicProperty);
         /// <summary> Remove a user-defined dynamic property</summary>
         public const string RemoveDynamicProperty = Common + nameof(RemoveDynamicProperty);
+        /// <summary> Assign an expression. </summary>
+        public const string AssignExpression = Common + nameof(AssignExpression);
     }
 }
